SRC/inference/requirements.txt
pyspark==3.3.2
kafka-python==2.0.2
mlflow==2.6.0
pandas==1.5.3
pyarrow==11.0.0
joblib==1.3.2
scikit-learn==1.2.2
gcsfs==2024.3.0
requests==2.31.0